pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "scikit-image>=0.25.2",
    "scipy>=1.16.1",
    "tensorflow>=2.14.0",
    "werkzeug>=3.1.3",
]
